use **kwargs for overriden AlternativeName.save()
Model.save() actually takes four keyword arguments. AlternativeName's overriden
save method was only implementing two of them, and this causes problems when
other parts of the code try to pass one of the missing ones (such as `using') to
this method.

Instead of adding the two missing keywords, it is cleaner to just pass the
keywords dict through with python's `**identifier' syntax. See for details:
https://docs.python.org/2/reference/compound_stmts.html#function-definitions