remove LanguageAltName
dld/model.py used the LanguageAltName class to uniquely bind together a
Language and an AlternativeName. This was confusing, and failed to accomplish
the intended purpose of maintainng the uniquenes of AlternativeNames. For
example, one could create two different AlternativeName objects, and specify
the same string value for their name fields. Because Django is implicitly
creating an auto-incrementing primary key for AlternativeNames, each of these
objects will be sufficently "unique" to pass the validation that
LanguageAltName is providing, yet will still have equivalent values.

The better method is to just specify `primary_key=True' for the AlternativeName
`name' field. This is the desired functionality, and accomplishes the same
purpose as the chunk of code in `lang_db.py' which was preventing duplicate
AlternativeName objects from being created.