CRITICAL: set DEBUG=False in langdeath/settings.py
When developing web-apps, the Django DEBUG mode is handy because it displays
debug information in the accessed webpage when errors occur. This information is
dangerously sensitive from a website security standpoint, and so DEBUG needs to
be disabled once a site is open to the world.

langdeath is primarily using Django for its database abstraction layer, not to
display a website, and so the security concerns posed by enabling DEBUG do not
apply.

Unfortunately, the other side effect of DEBUG mode is that it saves into memory
//every// SQL query made against the database while the program is running. As
langdeath makes a large number of database calls, this feature significantly and
unnecessarily increases the system memory required to run the software.