add check for empty needed_list
If the crossvalidation threshold is too high (or, if the crossvalidation results
are too low), it may be that no experiment passes the threshold, resulting in an
empty needed_list, and an empty self.best dataframe (no columns). This causes
problems with the apply aggregation functions, so I added a check to skip those
code paths when necessary.

Also, I added extra blank lines and a comment for increased readability.