Separate crossval scores from main dataframe
The csv-exported dataframe has includes a single extra row giving the
crossvalidation score of each experiment. This is useful for human reading
but is inconvenient for importing into analysis software because it makes the
string "crossval_res" appear to be an integrated language code. Better to deal
with column specific attributes separately.